refactor: Move jquants directory to sources layer
