+++ src/kabukit/sources/jquants/info.py
