+++ src/kabukit/sources/jquants/schema.py
