+++ tests/system/jquants/test_topix.py
@@ -2,7 +2,7 @@ import datetime
 
 import pytest
 
-from kabukit.jquants.client import JQuantsClient
+from kabukit.sources.jquants.client import JQuantsClient
 
 pytestmark = pytest.mark.system
